Gcc needs to be compiled with graphite in USE
It took me forever to figure out why gcc could not find isl even after installing different versions and recompiling gcc. I found by looking through solved issues that gcc needs to be compiled with graphite in the USE. Maybe it's obvious to more experienced people but I felt that something should be added specifying that this needs to be done since the error makes you go down the tree of thinking the issue is with isl. Feel free to reject this, I'm pretty new to Gentoo so I don't know what things people need to be told or not. It could also maybe be made a bit more obvious that there is chat group. I didn't notice the glitter bade for quite a while.