Suggest NTHREADS="auto" in documentation and samples
GCC 10 and higher support -flto=auto, allowing for gcc to determine
optimal number of cores to use. If -flto is set to a number, and certain
projects such as llvm are build, it will multiply this number by number
of make instances, causing too many processes to be created, ultimately
slowing down the build to a crawl on smaller devices with less ram(or
causing OOM conditions). This commit suggests using auto in configs for
ltoize and README.md.