Fix sys-config/ltoize to install symlinks at proper location in Prefix. (#862)
dosym seems to silently adds $EPREFIX to some parts of its arguments,
see https://bugs.gentoo.org/615594. This is a hackish way of having
Prefix install the symlinks at the expected location.

I updated the documentation to reflect that Prefix cannot use the
make.conf.lto symlink, at it sources an absolute path that is not
present in Prefix.